Conversão de Int para Long na obtenção de um Id em Foto.fromCursor()

+++ app/src/main/java/com/infact/nightour/model/Foto.java
@@ -47,7 +47,7 @@ public class Foto {
     public static Foto fromCursor(Cursor cursor) {
         Foto foto = new Foto();
 
-        foto.setId(cursor.getInt(cursor.getColumnIndexOrThrow(BD_ID)));
+        foto.setId(cursor.getLong(cursor.getColumnIndexOrThrow(BD_ID)));
         foto.setImagem(bitmapFromBlob(cursor.getBlob(cursor.getColumnIndexOrThrow(BD_IMAGEM))));
         foto.setDescricao(cursor.getString(cursor.getColumnIndexOrThrow(BD_DESCRICAO)));
         foto.setTimestamp(cursor.getLong(cursor.getColumnIndexOrThrow(BD_TIMESTAMP)));
